doyen_crewai/requirements.txt
crewai>=0.11.0
langchain>=0.1.0
chromadb>=0.4.22
lxml>=4.9.3
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-env>=1.1.1 